separated single grids component into control panel and grid system for better readability
