+++ src/app/grid-system/grid-system.component.css
@@ -1,0 +1,45 @@
+.grid {
+    display: grid;
+    gap: 1px;
+    max-width: 100%;
+    width: fit-content;
+    margin: 0 auto;
+}
+
+.grid-item {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    aspect-ratio: 1/1;
+    background-color: #f9f9f9;
+}
+
+/* Inner grid */
+.inner-grid {
+    display: grid;
+}
+
+/* Individual inner grid items */
+.inner-grid-item {
+    background-color: #e0e0e0;
+    border: 1px solid #bbbaba;
+}
+
+.svg-grid-container {
+    position: relative;
+    width: fit-content;
+    margin: 0 auto;
+}
+
+.svg-grid-container::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-image: url("data:image/svg+xml;base64,PHN2ZyBzdHlsZT0iY29sb3Itc2NoZW1lOiBsaWdodCBkYXJrOyIgd2lkdGg9IjE2IiBoZWlnaHQ9IjE2IiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPjxkZWZzPjxwYXR0ZXJuIGlkPSJncmlkIiB3aWR0aD0iMTYiIGhlaWdodD0iMTYiIHBhdHRlcm5Vbml0cz0idXNlclNwYWNlT25Vc2UiPjxwYXRoIGQ9Ik0gMCA0IEwgMTYgNCBNIDQgMCBMIDQgMTYgTSAwIDggTCAxNiA4IE0gOCAwIEwgOCAxNiBNIDAgMTIgTCAxNiAxMiBNIDEyIDAgTCAxMiAxNiIgZmlsbD0ibm9uZSIgc3R5bGU9InN0cm9rZTpsaWdodC1kYXJrKCNkMGQwZDAsICM0MjQyNDIpOyIgc3Ryb2tlPSIjZDBkMGQwIiBvcGFjaXR5PSIwLjIiIHN0cm9rZS13aWR0aD0iMSIvPjxwYXRoIGQ9Ik0gMTYgMCBMIDAgMCAwIDE2IiBmaWxsPSJub25lIiBzdHlsZT0ic3Ryb2tlOmxpZ2h0LWRhcmsoI2QwZDBkMCwgIzQyNDI0Mik7IiBzdHJva2U9IiNkMGQwZDAiIHN0cm9rZS13aWR0aD0iMSIvPjwvcGF0dGVybj48L2RlZnM+PHJlY3Qgd2lkdGg9IjEwMCUiIGhlaWdodD0iMTAwJSIgZmlsbD0idXJsKCNncmlkKSIvPjwvc3ZnPg==");
+    background-position: 0 0;
+    background-size: 16px 16px;
+    z-index: -1;
+}
