+++ src/app/grid-system/grid-system.component.html
@@ -1,0 +1,17 @@
+<div class="svg-grid-container">
+    <div class="grid" [style.grid-template-columns]="'repeat(' + outerGridSize + ', 1fr)'"
+         [style.grid-template-rows]="'repeat(' + outerGridSize + ', 1fr)'" [style.background]="selectedColor">
+      @for (item of items; track item) {
+        <div class="grid-item">
+          <div class="inner-grid"
+               [style.grid-template-columns]="'repeat(' + innerGridSize + ', ' + cellSize + 'px)'"
+               [style.grid-template-rows]="'repeat(' + innerGridSize + ', ' + cellSize + 'px)'">
+            @for (innerItem of innerGridItems; track innerItem) {
+              <div class="inner-grid-item"></div>
+            }
+          </div>
+        </div>
+      }
+    </div>
+  </div>
+
