+++ src/app/grid-system/grid-system.component.ts
@@ -1,0 +1,31 @@
+import { Component, Input, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-grid-system',
+  templateUrl: './grid-system.component.html',
+  styleUrls: ['./grid-system.component.css']
+})
+export class GridSystemComponent implements OnInit {
+  @Input() outerGridSize: number = 20;
+  @Input() innerGridSize: number = 4;
+  @Input() cellSize: number = 13;
+  @Input() selectedColor: string = '#8f8e8e';
+
+  items: number[] = [];
+  innerGridItems: number[] = [];
+
+  ngOnInit() {
+    this.generateGrid();
+    this.generateInnerGridItems();
+  }
+
+  generateGrid() {
+    const totalBlocks = this.outerGridSize * this.outerGridSize;
+    this.items = Array.from({ length: totalBlocks }, (_, i) => i + 1);
+  }
+
+  generateInnerGridItems() {
+    const totalInnerItems = this.innerGridSize * this.innerGridSize;
+    this.innerGridItems = Array.from({ length: totalInnerItems }, (_, i) => i + 1);
+  }
+}
