+++ src/app/material.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
+
+@NgModule({
+  exports: [MatInputModule, MatFormFieldModule, MatIconModule, MatButtonModule, MatSelectModule]
+})
+export class MaterialModule {}
