+++ src/app/sidebar/sidebar.component.html
@@ -10,7 +10,7 @@
         </button>
         <span><i>WorkFlow Editor</i></span>
       </mat-toolbar>
-  
+
       <div class="content">
        
         <app-grids></app-grids>
